refactor: Background 컴포넌트 폴더 경로 수정
