+++ client/src/components/Background/index.tsx
